requirements.txt
mysql-connector==2.2.9
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
